invoices are working now

+++ src/main/java/org/tvolkov/controller/InvoiceController.java
@@ -13,26 +13,26 @@ public class InvoiceController {
     private InvoiceService invoiceService;
 
     @GetMapping(params = "customerId, month")
-    public String getInvoicesPerMonth(@RequestParam("customerId") long customerId,
+    public String getInvoicesPerMonth(@RequestParam("customerId") int customerId,
                                       @RequestParam("month") int month){
         return invoiceService.getInvoicesPerMonth(customerId, month, "all");
     }
 
     @GetMapping(params = "customerId, filter, month")
-    public String getShopInvoicesPerMonth(@RequestParam("customerId") long customerId,
+    public String getShopInvoicesPerMonth(@RequestParam("customerId") int customerId,
                                           @RequestParam("filter") String filter,
                                           @RequestParam("month") int month){
         return invoiceService.getInvoicesPerMonth(customerId, month, filter);
     }
 
     @GetMapping( params = "customerId, addressId")
-    public String getInvoicesHistoryPerAddress(@RequestParam("customerId") long customerId,
+    public String getInvoicesHistoryPerAddress(@RequestParam("customerId") int customerId,
                                                @RequestParam("addressId") String addressId){
         return invoiceService.getInvoicesPerAddress(customerId, addressId);
     }
 
     @GetMapping(params = "customerId")
-    public String getFullInvoicesHistory(@RequestParam("customerId") long customerId){
+    public String getFullInvoicesHistory(@RequestParam("customerId") int customerId){
         return invoiceService.getFullInvoicesHistory(customerId);
     }
