+++ src/main/java/org/tvolkov/repository/InvoiceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import org.tvolkov.model.Invoice;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
 
 }
